Reuse same cypress tests for on demand radio and TV
