--- cypress/integration/pages/onDemandRadio/testsForCanonicalOnly.js
@@ -1,40 +1,0 @@
-import appConfig from '../../../../src/server/utilities/serviceConfigs';
-import envConfig from '../../../support/config/envs';
-import { getEmbedUrl, isExpired, dataEndpointOverride } from './helpers';
-
-export default ({ service, pageType, variant }) => {
-  describe(`testsForCanonicalOnly for ${service} ${pageType}`, () => {
-    describe('Audio Player', () => {
-      const { lang } = appConfig[service][variant];
-      let embedUrl;
-      let isExpiredEpisode;
-
-      it('should render an iframe with a valid URL', () => {
-        cy.request(
-          `${Cypress.env('currentPath')}.json${dataEndpointOverride()}`,
-        ).then(({ body: jsonData }) => {
-          embedUrl = getEmbedUrl(jsonData, lang);
-          isExpiredEpisode = isExpired(jsonData);
-
-          if (!isExpiredEpisode) {
-            cy.get(`iframe[src*="${embedUrl}"]`).should('be.visible');
-            cy.testResponseCodeAndType(embedUrl, 200, 'text/html');
-          } else {
-            cy.log(`Episode is expired: ${Cypress.env('currentPath')}`);
-          }
-        });
-      });
-    });
-
-    describe('Chartbeat', () => {
-      if (envConfig.chartbeatEnabled) {
-        it('should have a script with src value set to chartbeat source', () => {
-          cy.hasScriptWithChartbeatSrc();
-        });
-        it('should have chartbeat config set to window object', () => {
-          cy.hasGlobalChartbeatConfig();
-        });
-      }
-    });
-  });
-};
